safety net when popping operands
If the size of the stack is `0`, then the following commands will become nops: `p+-*_id`.
If the size of the stack is `1`, then the following commands will become nops: `+-*_`.
I have fixed the code for `p`. It reset the ToS after popping to `0` before I fixed it. Now, we don't set anything to zero, and just decrement the size, as is done in the codes for `+-*_`.